Add files via upload
here I have only posted the code for the game to work and later on I will update it with the resources I used. As you can see I used classes to create my units and my buildings an i imported them in the main.py in order to have a clearer code. The code has some flaws and could be better but it might work as a good template for other games as well.